moves loader to folder
