+++ src/index.js
@@ -1,7 +1,5 @@
-import _ from './js/loader';
-import _ from './js/categories';
+import './js/categories';
 import './js/toggle-switch';
-import './js/loader';
 import './js/best-sellers';
 import './js/support-slider';
 import './js/auth/auth-form-base';
