+++ src/js/loader/_loader.scss
